[#9] Chore: DependencyName 뎁스 단축 및 불필요 모듈 삭제
