--- Projects/Data/Booking/Project.swift
@@ -1,27 +1,0 @@
-//
-//  BookingDataProject.swift
-//  Booking
-//
-//  Created by  on .
-//
-
-
-import ProjectDescription
-import ProjectDescriptionHelpers
-
-let project = Project(
-    name: "BookingData",
-    settings: .defaultSettings,
-    targets: [
-        .makeDynamicFrameworkTarget(
-            name: "BookingData",
-            deploymentTargetsVersion: "\(Constants.targetVersion)",
-            dependencies: [
-                DependencyContainer.BookingDomain,
-                DependencyContainer.HGNetwork,
-                DependencyContainer.HGDataBase,
-            ],
-            hasResources: false
-        )
-    ]
-)
