--- Projects/Data/Booking/Sources/BookingAPI.swift
@@ -1,14 +1,0 @@
-//
-//  BookingAPI.swift
-//  Booking
-//
-//  Created by  on .
-//
-
-import Foundation
-import HGNetwork
-import BookingDomain
-
-struct BookingAPI {
-
-}
