--- Projects/Data/Chat/Sources/ChatAPI.swift
@@ -1,14 +1,0 @@
-//
-//  ChatAPI.swift
-//  Chat
-//
-//  Created by  on .
-//
-
-import Foundation
-import HGNetwork
-import ChatDomain
-
-struct ChatAPI {
-
-}
