--- Projects/Data/User/Sources/UserAPI.swift
@@ -1,14 +1,0 @@
-//
-//  UserAPI.swift
-//  User
-//
-//  Created by  on .
-//
-
-import Foundation
-import HGNetwork
-import UserDomain
-
-struct UserAPI {
-
-}
