--- Projects/Domain/Booking/Sources/BookingRepository.swift
@@ -1,12 +1,0 @@
-//
-//  BookingRepository.swift
-//  Booking
-//
-//  Created by  on .
-//
-
-import Foundation
-
-public protocol BookingRepository {
-
-}
