--- Projects/Domain/Chat/Sources/ChatRepository.swift
@@ -1,12 +1,0 @@
-//
-//  ChatRepository.swift
-//  Chat
-//
-//  Created by  on .
-//
-
-import Foundation
-
-public protocol ChatRepository {
-
-}
