--- Projects/Domain/User/Sources/UserRepository.swift
@@ -1,12 +1,0 @@
-//
-//  UserRepository.swift
-//  User
-//
-//  Created by  on .
-//
-
-import Foundation
-
-public protocol UserRepository {
-
-}
