--- Projects/Features/Booking/Sources/BookingView.swift
@@ -1,9 +1,0 @@
-//
-//  BookingView.swift
-//  Booking
-//
-//  Created by  on .
-//
-
-import SwifUI
-import BookingDomain
