--- Projects/Features/Chat/Sources/ChatView.swift
@@ -1,16 +1,0 @@
-//
-//  ChatView.swift
-//  Chat
-//
-//  Created by  on .
-//
-
-import SwiftUI
-import ChatDomain
-import HGDesignSystem
-
-struct ChatView: View {
-    var body: some View {
-        Text("")
-    }
-}
