--- Projects/Features/User/Sources/UserView.swift
@@ -1,9 +1,0 @@
-//
-//  UserView.swift
-//  User
-//
-//  Created by  on .
-//
-
-import SwifUI
-import UserDomain
