+++ Tuist/ProjectDescriptionHelpers/DependencyName.swift
@@ -1,0 +1,26 @@
+//
+//  DependencyName.swift
+//  ProjectDescriptionHelpers
+//
+//  Created by iOS신상우 on 5/4/25.
+//
+
+import Foundation
+import ProjectDescription
+
+public enum DependencyName: String {
+    case onboarding = "Onboarding"
+    case home = "Home"
+    case reservation = "Reservation"
+    case reservationHistory = "ReservationHistory"
+    case createReservation = "CreateReservation"
+    case myPage = "MyPage"
+    
+    case hgNetwork = "HGNetwork"
+    case hgDatabase = "HGDataBase"
+    case hgLogger = "HGLogger"
+    case hgCommon = "HGCommon"
+    case hgThridParty = "HGThridParty"
+    
+    case hgDesignSystem = "HGDesignSystem"
+}
